requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==5.3.1
    # via kombu
anyio==4.12.1
    # via watchfiles
asgiref==3.11.0
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   django-cors-headers
asn1crypto==1.5.1
    # via webauthn
attrs==25.4.0
    # via
    #   service-identity
    #   twisted
autobahn==25.12.2
    # via daphne
automat==25.4.16
    # via twisted
billiard==4.2.4
    # via celery
black==25.12.0
    # via -r requirements.in
cbor2==5.8.0
    # via
    #   autobahn
    #   webauthn
celery==5.5.3
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-celery-results
certifi==2026.1.4
    # via requests
cffi==2.0.0
    # via
    #   autobahn
    #   cryptography
channels==4.3.2
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.3.0
    # via -r requirements.in
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
constantly==23.10.4
    # via twisted
cryptography==46.0.3
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
    #   webauthn
daphne==4.2.1
    # via -r requirements.in
django==6.0.1
    # via
    #   -r requirements.in
    #   channels
    #   django-appconf
    #   django-celery-beat
    #   django-celery-results
    #   django-cors-headers
    #   django-extensions
    #   django-polymorphic
    #   django-silk
    #   django-sslserver
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
django-appconf==1.2.0
    # via
    #   django-imagekit
    #   django-webauthn
django-celery-beat==2.1.0
    # via -r requirements.in
django-celery-results==2.6.0
    # via -r requirements.in
django-cors-headers==4.9.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-extensions==4.1
    # via -r requirements.in
django-imagekit==6.0.0
    # via -r requirements.in
django-polymorphic==4.1.0
    # via -r requirements.in
django-silk==5.4.3
    # via -r requirements.in
django-sslserver==0.22
    # via -r requirements.in
django-timezone-field==4.2.3
    # via django-celery-beat
django-webauthn==0.0.2
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
gprof2dot==2025.4.14
    # via django-silk
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.11
    # via
    #   anyio
    #   hyperlink
    #   requests
    #   twisted
incremental==24.11.0
    # via twisted
kombu==5.5.4
    # via celery
msgpack==1.1.2
    # via
    #   autobahn
    #   channels-redis
mypy-extensions==1.1.0
    # via black
packaging==25.0
    # via
    #   black
    #   incremental
    #   kombu
pathspec==1.0.3
    # via black
pilkit==3.0
    # via django-imagekit
pillow==12.1.0
    # via
    #   -r requirements.in
    #   pilkit
platformdirs==4.5.1
    # via black
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.11
    # via -r requirements.in
py-ubjson==0.16.1
    # via autobahn
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.23
    # via cffi
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pyopenssl==25.3.0
    # via
    #   twisted
    #   webauthn
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via celery
pytokens==0.3.0
    # via black
pytz==2025.2
    # via django-timezone-field
redis==7.1.0
    # via
    #   -r requirements.in
    #   channels-redis
requests==2.32.5
    # via -r requirements.in
service-identity==24.2.0
    # via twisted
six==1.17.0
    # via python-dateutil
sqlparse==0.5.5
    # via
    #   django
    #   django-silk
twisted[tls]==25.5.0
    # via daphne
txaio==25.12.2
    # via autobahn
typing-extensions==4.15.0
    # via twisted
tzdata==2025.3
    # via kombu
ujson==5.11.0
    # via autobahn
urllib3==2.5.0
    # via
    #   -r requirements.in
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==1.1.1
    # via -r requirements.in
wcwidth==0.2.14
    # via prompt-toolkit
webauthn==2.7.0
    # via -r requirements.in
zope-interface==8.1.1
    # via twisted
